v1.23.0 additions, changes, fixes
- Add back mass of 1 to Ace_Banana now that it is a useful item in AceX Field Rations
- Add icons from 2PzD mission tempalte to this mod in order to reduce misson file size
- Add new custom recoils for small arms
This completely redoes recoil for all small arms, providing a much better balance
- Change indirectHitRange of small german aircraft bombs to cause a less splash damage
- Change mass and load of IFA flamethrower magazine and backpacks
- Change more FOW magazines to use IFA ammo classes for improved damage compatibility
- Change several FOW and LEN magazines to scope = 1 so fewer duplicate magazines appear in the arsenal
The magazines are still useable if spawned by script or picked up off AI or from boxes/vehicles
Arma doesn't handle flamethrowers well, since in Arma the fuel is a 'magazine' which is loaded in the flamethrower weapon.
This means that the flamethrower backpack has tons of empty space that can be abused for extra carrying capacity since it has no fuel magazine in it.
Now the fuel backpack always has the mass of an empty tank, while the fuel magazine has the weight of the fuel.
There is also a new backpack class that has no storage space. So add the `B_LIB_US_M2Flamethrower` backpack, add the weapon and magazine, then remove that backpack and add `B_LIB_US_M2Flamethrower_noLoad` instead.
- Change the order of config entries for classes so that displayName is highest for easier readibility
- Change timeToLive of IFA flamethrower projectiles so they don't go farther than the flame effect
- Fix several class inheritance after a complete review of all changed classes
- Fix some line endings
- Remove duplicates of some classes that were accidentally declared twice